+++ src/components/Embeds/AirGauge.js
@@ -4,6 +4,8 @@ import PropTypes from 'prop-types';
 import Grid from '@material-ui/core/Grid';
 import CityGauge from '../Header/JumbotronContent/AirCityHeaderContent/CityGauge';
 
+import { API, formatCurrentP2Stats } from '../../api';
+
 class AirGauge extends React.Component {
   constructor(props) {
     super(props);
@@ -11,7 +13,7 @@ class AirGauge extends React.Component {
     this.state = {
       stats: {
         average: '--',
-        description: ''
+        averageDescription: ''
       }
     };
   }
@@ -21,45 +23,20 @@ class AirGauge extends React.Component {
     const params = new URLSearchParams(location.search);
     const city = params.get('city');
 
-    const formatCurrentP2Stats = (data, isPm2 = false) => {
-      const formatted = {};
-      ['average', 'maximum', 'minimum'].forEach(stat => {
-        const parsed = Number.parseFloat(data[stat]);
-        if (isPm2 && stat === 'average') {
-          formatted.description = `measurements not recorded`;
-          if (!Number.isNaN(parsed)) {
-            let difference = 25.0 - parsed;
-            let position = 'below';
-            if (parsed > 25.0) {
-              difference = parsed - 25.0;
-              position = 'above';
-            }
-            const percentage = ((difference / 25.0) * 100).toFixed(2);
-            formatted.description = `${percentage}% ${position} the safe level`;
-          }
-        }
-        formatted[stat] = Number.isNaN(parsed) ? '--' : parsed.toFixed(2);
-      });
-      return formatted;
-    };
+    this.setState({
+      stats: {
+        average: '--',
+        averageDescription: 'loading'
+      }
+    });
 
-    const processJson = json => {
+    API.getCurrentAirData(city, json => {
       let stats = {};
       if (json.count === 1) {
         stats = formatCurrentP2Stats(json.results[0].P2 || {}, true);
       }
       this.setState({ stats });
-    };
-
-    this.setState({
-      stats: {
-        average: '--',
-        description: 'loading'
-      }
     });
-    fetch(`https://api.sensors.africa/v2/data/air/?city=${city}`)
-      .then(data => data.json())
-      .then(json => processJson(json));
   }
 
   render() {
@@ -75,7 +52,7 @@ class AirGauge extends React.Component {
       >
         <CityGauge
           airPollMeasurement={stats.average}
-          airPollDescription={stats.description}
+          airPollDescription={stats.averageDescription}
         />
       </Grid>
     );
