+++ src/components/Embeds/AirMap.js
@@ -1,19 +1,23 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const CITIES_LOCATION = {
-  nairobi: '12/-1.2709/36.8169',
-  lagos: '12/6.4552/3.4198',
-  'dar-es-salaam': '12/-6.8555/39.1518'
-};
+import { CITIES_LOCATION } from '../../api';
 
 function AirMap({ location }) {
   const params = new URLSearchParams(location.search);
   const city = params.get('city');
 
-  window.location = `https://map.aq.sensors.africa/#${CITIES_LOCATION[city]}`;
-
-  return <React.Fragment />;
+  return (
+    <iframe
+      title="Sensor Map"
+      src={`https://map.aq.sensors.africa/#${CITIES_LOCATION[city].location}`}
+      style={{
+        height: '100vh',
+        width: '100vw'
+      }}
+      frameBorder="none"
+    />
+  );
 }
 
 AirMap.propTypes = {
